% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{pass_rate_calibration-methods}
\alias{pass_rate_calibration-methods}
\alias{print.pass_rate_calibration}
\alias{summary.pass_rate_calibration}
\alias{print.summary_pass_rate_calibration}
\alias{plot.pass_rate_calibration}
\title{Methods for pass_rate_calibration objects}
\usage{
\method{print}{pass_rate_calibration}(x, digits = 4, ...)

\method{summary}{pass_rate_calibration}(object, top = 10L, ...)

\method{print}{summary_pass_rate_calibration}(x, digits = 4, ...)

\method{plot}{pass_rate_calibration}(x, type = c("target_error", "multipliers"), top = 20L, ...)
}
\arguments{
\item{x, object}{A \code{pass_rate_calibration} object (or, for the summary
printer, a \code{summary_pass_rate_calibration} object).}

\item{digits}{Number of decimal places to display.}

\item{...}{Further arguments passed to the underlying print or graphics
functions.}

\item{top}{Number of largest-error targets to display.}

\item{type}{Plot type: \code{"target_error"} for target errors or
\code{"multipliers"} for the weight-multiplier histogram.}
}
\value{
\code{print} and \code{plot} return their input invisibly; \code{summary} returns a
\code{summary_pass_rate_calibration} object.
}
\description{
S3 methods that print, summarize and plot calibration results.
}
