% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/replicate.R
\name{replicate_variance}
\alias{replicate_variance}
\title{Replicate-weight variance of a calibrated estimate}
\usage{
replicate_variance(object, x, statistic = c("total", "mean"))
}
\arguments{
\item{object}{A \code{replicate_calibration} object from
\code{\link[=calibrate_replicate_weights]{calibrate_replicate_weights()}}.}

\item{x}{Numeric study variable, one value per observation.}

\item{statistic}{Either \code{"total"} (weighted sum) or \code{"mean"} (weighted mean).}
}
\value{
A list with \code{estimate}, \code{variance} and \code{se}.
}
\description{
Computes the point estimate, variance and standard error of a weighted total
or mean of a study variable, using calibrated replicate weights.
}
\examples{
d <- example_rate_data(300)
fit <- calibrate_rates(d, "qualified", "initial_weight",
                       groups = list(sex = c(M = 0.72, F = 0.68)))
repw <- d$initial_weight * matrix(stats::runif(nrow(d) * 5, 0.8, 1.2), ncol = 5)
rc <- calibrate_replicate_weights(fit, repw)
replicate_variance(rc, d$qualified, statistic = "mean")
}
